requirements.txt
fastapi
uvicorn
paho-mqtt
supabase
python-dotenv